refactor
- userId Long -> Integer 변경
- userId @PathVariable 방식으로 임시 설정
- @JsonNaming 추가

+++ src/main/java/com/bovo/Bovo/modules/archive/repository/MyBooksRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface MyBooksRepository extends JpaRepository<MyBooks, Long> {
     // userId 컬럼(또는 연관된 User 엔티티의 id)으로 등록된 책 목록을 조회
-    List<MyBooks> findByUserId(Long userId);
+    List<MyBooks> findByUserId(Integer userId);
 }
