+++ src/main/java/com/bovo/Bovo/modules/archive/service/MyBooksService.java
@@ -3,6 +3,5 @@ package com.bovo.Bovo.modules.archive.service;
 import com.bovo.Bovo.modules.archive.dto.Response.MyBooksResponse;
 
 public interface MyBooksService {
-    MyBooksResponse getMyBooks(Long userId);
+    MyBooksResponse getMyBooks(Integer userId);
 }
-
